+++ docs/changelog.md
@@ -19,6 +19,9 @@ Requires libvips v8.13.1
 * Ensure AVIF output is always 8-bit.
   [#3358](https://github.com/lovell/sharp/issues/3358)
 
+* Ensure greyscale images can be trimmed (regression in 0.31.0).
+  [#3386](https://github.com/lovell/sharp/issues/3386)
+
 ### v0.31.0 - 5th September 2022
 
 * Drop support for Node.js 12, now requires Node.js >= 14.15.0.